+++ src/extension.ts
@@ -1,13 +1,14 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 import { BranchSwitcher } from './tree/scm/branches';
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
+let branchSwitcher: BranchSwitcher;
+
 export function activate(context: vscode.ExtensionContext) {
-    new BranchSwitcher(context);
+    branchSwitcher = new BranchSwitcher(context);
 }
 
-// this method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() {
+    if (branchSwitcher) {
+        branchSwitcher.dispose();
+    }
+}