+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.2.0] - 2019-10-XX
+
+* Add ability to create a new branch from current branch
+* Add confirmations for deleting branches
+* Added unit tests
+
 ## [1.1.1] - 2019-09-25
 
 * Fix required module not getting loaded, preventing startup